refactor MethodTypingContext#adaptMethod moved to ClassTypingContext
